chore: added apps/installation routes to future
