+++ turbo.json
@@ -220,6 +220,7 @@
     "APP_ROUTER_APPS_CATEGORIES_ENABLED",
     "APP_ROUTER_APPS_ENABLED",
     "APP_ROUTER_APPS_INSTALLED_CATEGORY_ENABLED",
+    "APP_ROUTER_APPS_INSTALLATION_ENABLED",
     "APP_ROUTER_APPS_SLUG_ENABLED",
     "APP_ROUTER_APPS_SLUG_SETUP_ENABLED",
     "APP_ROUTER_BOOKINGS_STATUS_ENABLED",
